backend/requirements.txt
Django>=5.1.5
djangorestframework>=3.15.0
django-cors-headers>=4.6.0
Pillow>=11.0.0
